Store configuration files in a subdirectory
The is preferred due to the better security properties - the directory may be made writable without allowing modifications to the binary files. However, since this breaks compatibility with current installations, it is only used if the
directory exists. Otherwise the original behavior (config files directly next to the main binary files) is used.